refactor(overlayfs): split implementation into modules
Split the overlayfs implementation across focused internal modules for filesystem setup, inode operations, file and directory operations, lookup, readdir, copy-up, rename, whiteout handling, workdir helpers, and path helpers.

Keep the existing behavior and public API surface intact while reducing mod.rs to a small module entry point.
